feat: 영화 상세 정보 처리 로직 분리 및 배너에서 버튼 클릭 이벤트 연결
