+++ src/view/Banner.ts
@@ -2,16 +2,17 @@ import createDOMElement from '../util/createDomElement';
 import Button from '../component/Button';
 import { IMAGE_BASE_URL } from '../constant';
 import { MoveType } from '../type';
+import { handleMovieDetail } from './events/handleMovieDetail';
 
 function Banner({ movie }: { movie: MoveType }) {
-  const { backdropPath, voteAverage, title } = movie;
+  const { backdropPath, voteAverage, title, id } = movie;
   return createDOMElement({
     tag: 'header',
     children: [
       createDOMElement({
         tag: 'div',
         className: 'background-container',
-        children: [BackDrop({ backDropUrl: backdropPath }), TopRatedMovie({ voteAverage, title })]
+        children: [BackDrop({ backDropUrl: backdropPath }), TopRatedMovie({ voteAverage, title, id })]
       })
     ]
   });
@@ -31,7 +32,7 @@ function BackDrop({ backDropUrl }: { backDropUrl: string | null }) {
   });
 }
 
-function TopRatedMovie({ voteAverage, title }: { voteAverage: number; title: string }) {
+function TopRatedMovie({ voteAverage, title, id }: { voteAverage: number; title: string; id: number }) {
   return createDOMElement({
     tag: 'div',
     className: 'top-rated-container',
@@ -61,7 +62,12 @@ function TopRatedMovie({ voteAverage, title }: { voteAverage: number; title: str
             className: 'title',
             textContent: title
           }),
-          Button({ text: '자세히보기', id: 'bannerMovieButton', className: 'primary' })
+          Button({
+            text: '자세히보기',
+            id: 'bannerMovieButton',
+            className: 'primary',
+            onClick: () => handleMovieDetail(id)
+          })
         ]
       })
     ]
